fix(NPE): return a string instead of a object for worldPath
As initially written worldPath was a file object which resulted in a NPE being thrown even though the path was being returned.

This commit makes worldPath return a string instead of an object. The NPE is no longer thrown. I believe this will still allow the code to function as intended.